+++ src/main/app/.eslintrc
@@ -19,7 +19,8 @@
         "@typescript-eslint",
         "prettier",
         "lodash",
-        "import"
+        "import",
+        "react-refresh"
     ],
     "env": {
         "browser": true,
@@ -27,6 +28,13 @@
     },
     "rules": {
         "playwright/expect-expect": "off",
+        "react-refresh/only-export-components": [
+            "warn",
+            {
+                "allowConstantExport": true,
+                "checkJS": true
+            }
+        ],
         "react/prop-types": "off",
         "react/display-name": "off",
         "@typescript-eslint/no-empty-function": "off",
