+++ src/main/app/package.json
@@ -84,6 +84,7 @@
     "eslint-plugin-playwright": "^0.15.1",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-refresh": "^0.4.1",
     "husky": "^8.0.3",
     "jsdom": "^22.1.0",
     "lint-staged": "^13.2.2",
