+++ src/main/app/src/components/haku/hakutulos/hakutulosKortit/KoulutusKortti.tsx
@@ -6,10 +6,8 @@ import {
   TimelapseOutlined,
   HomeWorkOutlined,
 } from '@mui/icons-material';
-import { TFunction } from 'i18next';
 import { isEmpty } from 'lodash';
 import { useTranslation } from 'react-i18next';
-import { match, P } from 'ts-pattern';
 
 import { EntiteettiKortti } from '#/src/components/common/EntiteettiKortti';
 import { KoulutusKorttiLogo } from '#/src/components/common/KorttiLogo';
@@ -18,6 +16,8 @@ import { useVisibleKoulutustyyppi } from '#/src/hooks/useVisibleKoulutustyyppi';
 import { localize } from '#/src/tools/localization';
 import { getLocalizedKoulutusLaajuus } from '#/src/tools/utils';
 import { ToteutustenTarjoajat, Translateable } from '#/src/types/common';
+
+import { getToteutustenTarjoajat } from './getToteutustenTarjoajat';
 
 type Props = {
   koulutus: {
@@ -30,24 +30,6 @@ type Props = {
     isAvoinKorkeakoulutus: boolean;
   };
   isSmall?: boolean;
-};
-
-export const getToteutustenTarjoajat = (
-  t: TFunction,
-  toteutustenTarjoajat?: ToteutustenTarjoajat
-) =>
-  match(toteutustenTarjoajat)
-    .with({ count: 0 }, () => t('haku.ei-koulutuksen-tarjoajia'))
-    .with({ count: 1 }, () => localize(toteutustenTarjoajat?.nimi))
-    .with(
-      { count: P.number },
-      () => `${toteutustenTarjoajat?.count} ${t('haku.koulutuksen-tarjoajaa')}`
-    )
-    .otherwise(() => undefined);
-
-const useToteutustenTarjoajat = (toteutustenTarjoajat?: ToteutustenTarjoajat) => {
-  const { t } = useTranslation();
-  return getToteutustenTarjoajat(t, toteutustenTarjoajat);
 };
 
 export const KoulutusKortti = ({ koulutus, isSmall }: Props) => {
@@ -69,7 +51,7 @@ export const KoulutusKortti = ({ koulutus, isSmall }: Props) => {
       .replace(/\.,<\/li>/gm, '.</li>')
       .replace(/<[^>]*>/gm, '') || t('haku.ei_kuvausta');
 
-  const toteutustenTarjoajatText = useToteutustenTarjoajat(toteutustenTarjoajat);
+  const toteutustenTarjoajatText = getToteutustenTarjoajat(t, toteutustenTarjoajat);
 
   const koulutustyyppiText = useVisibleKoulutustyyppi({
     koulutustyyppi,
